todo 600

+++ playground/leetcode-440--2018-09-01-153016/snippet.cpp
@@ -41,35 +41,26 @@ class Solution {
 public:
     int findKthNumber(int n, int k) {
         int result = 1;
-        for(--k; k > 0; )
-        {
+        for (--k; k > 0;) {
             // calculate #|{result, result*, result**, result***, ...}|
-            int count = 0;
-            for (long long first = static_cast<long long>(result), last = first + 1;
-                first <= n; // the interval is not empty
-                first *= 10, last *= 10) // increase a digit
-            {
+            long count = 0;
+            for (long first = result, last = first + 1; first <= n; first *= 10, last *= 10) {
                 // valid interval = [first, last) union [first, n]
-                count += static_cast<int>((min(n + 1LL, last) - first)); // add the length of interval
+                // add the length of interval
+                count += min(n + 1l, last) - first;
             }
-            
-            if (k >= count)
-            {   // skip {result, result*, result**, result***, ...}
+            if (k >= count) { // skip {result, result*, result**, result***, ...}
                 // increase the current prefix
                 ++result;
                 k -= count;
-            }
-            else
-            {   // not able to skip all of {result, result*, result**, result***, ...}
+            } else { // not able to skip all of {result, result*, result**, result***, ...}
                 // search more detailedly
                 result *= 10;
                 --k;
             }
         }
-        return result;       
+        return result;
     }
 };
 
-int mymain(int argc, char *argv[]) {
-    return 0;
-}
+int mymain(int argc, char* argv[]) { return 0; }
